% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{connect}
\alias{connect}
\title{Connect to an OmicsDS workspace/array}
\usage{
connect(workspace, array)
}
\arguments{
\item{workspace}{path to an OmicsDS workspace}

\item{array}{name of the array in the OmicsDS workspace}
}
\value{
handle representation of the OmicsDS instance for the given workspace and array
}
\description{
Get a handle to an OmicsDS instance given a workspace/array
Calling disconnect will release resources held
by the native OmicsDS.
}
\examples{
\dontrun{
omicsds_handle <- omicsds::connect(workspace="~/feature-level-ws", array="18K_samples")
}

}
