% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{disconnect}
\alias{disconnect}
\title{Disconnect from an OmicsDS workspace/array}
\usage{
disconnect(handle)
}
\arguments{
\item{handle}{OmicsDS instance representation for a workspace/array}
}
\description{
Disconnect from a given OmicsDS instance to release resources held by the
instance.
}
\examples{
\dontrun{
omicsds_handle <- omicsds::connect(workspace="~/feature-level-ws", array="18K_samples")
omicsds::disconnect(omicsds_handle)
}

}
