\name{omicsds-package}
\alias{omicsds-package}
\alias{omicsds}
\docType{package}
\title{
  OmicsDS R Bindings
}
\description{
  R Interface via bindings to the native OmicsDS library
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
OmicsDS OpenSource
Maintainer:  OmicsDS<nalinigans@github.com>
}
\references{
        github:OmicsDataAutomation/OmicsDS
}
\examples{
  \dontrun{
     initialize()
     query()
  }
}
