% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{query_features}
\alias{query_features}
\title{Slice OmicsDS array by features/samples}
\usage{
query_features(handle, features = NULL, sample_range = NULL)
}
\arguments{
\item{handle}{OmicsDS instance representation for a workspace/array}

\item{features}{slice by list of features or NULL for all features}

\item{sample_range}{slice by sample range or NULL for all samples}
}
\value{
R data.frame for the OmicsDS workspace array sliced by features/samples
}
\description{
Get a slice as a DataFrame from an OmicsDS instance referenced by handle,
given features and a sample range. Features and sample range can be
NULL, in which case all features and/or samples are returned from
OmicsDS.
}
\examples{
\dontrun{
omicsds_handle <- omicsds::connect(workspace="~/feature-level-ws", array="18K_samples")
df <- omicsds::query_features(handle=omicsds_handle, features=c("ENSG00000138190", "ENSG00000243485"), sample_range=c(0, 2))
df1 <- omicsds::query_features(handle=omicsds_handle, features=NULL, sample_range=c(0, 2))
df2 <- omicsds::query_features(handle=omicsds_handle, features=c("ENSG00000138190"), sample_range=NULL)
df3 <- omicsds::query_features(handle=omicsds_handle, features=NULL, sample_range=NULL)
}
}
