% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omicsds-package.R
\docType{package}
\name{omicsds-package}
\alias{omicsds-package}
\title{R bindings to the native OmicsDS Library}
\description{
This package features experimental R Bindings to the native OmicsDS library.
}
